Proyecto2/consumer/requirements.txt
paho-mqtt==2.1.0
pymongo==4.8.0
python-dotenv==1.0.0